# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "查课管理系统 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/academic/semesters": {
      "post": {
        "tags": [
          "academic"
        ],
        "summary": "Create Semester",
        "operationId": "create_semester_api_v1_academic_semesters_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SemesterCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Semester Api V1 Academic Semesters Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "academic"
        ],
        "summary": "List Semesters",
        "operationId": "list_semesters_api_v1_academic_semesters_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 16
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码，从 1 开始",
              "default": 1,
              "title": "Page"
            },
            "description": "页码，从 1 开始"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "每页条数",
              "default": 20,
              "title": "Page Size"
            },
            "description": "每页条数"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Semesters Api V1 Academic Semesters Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/academic/semesters/{semester_id}": {
      "patch": {
        "tags": [
          "academic"
        ],
        "summary": "Update Semester",
        "operationId": "update_semester_api_v1_academic_semesters__semester_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "semester_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Semester Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SemesterUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Semester Api V1 Academic Semesters  Semester Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "academic"
        ],
        "summary": "Get Semester",
        "operationId": "get_semester_api_v1_academic_semesters__semester_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "semester_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Semester Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Semester Api V1 Academic Semesters  Semester Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/academic/semesters/{semester_id}/period-definitions/{period_no}": {
      "put": {
        "tags": [
          "academic"
        ],
        "summary": "Upsert Period Definition",
        "operationId": "upsert_period_definition_api_v1_academic_semesters__semester_id__period_definitions__period_no__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "semester_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Semester Id"
            }
          },
          {
            "name": "period_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Period No"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeriodDefinitionUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Upsert Period Definition Api V1 Academic Semesters  Semester Id  Period Definitions  Period No  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/academic/semesters/{semester_id}/period-definitions": {
      "get": {
        "tags": [
          "academic"
        ],
        "summary": "List Period Definitions",
        "operationId": "list_period_definitions_api_v1_academic_semesters__semester_id__period_definitions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "semester_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Semester Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Period Definitions Api V1 Academic Semesters  Semester Id  Period Definitions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/academic/semesters/{semester_id}/period-definitions/{period_id}": {
      "delete": {
        "tags": [
          "academic"
        ],
        "summary": "Delete Period Definition",
        "operationId": "delete_period_definition_api_v1_academic_semesters__semester_id__period_definitions__period_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "semester_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Semester Id"
            }
          },
          {
            "name": "period_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Period Id"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/academic/semesters/{semester_id}/calendar-overrides": {
      "post": {
        "tags": [
          "academic"
        ],
        "summary": "Create Calendar Override",
        "operationId": "create_calendar_override_api_v1_academic_semesters__semester_id__calendar_overrides_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "semester_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Semester Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarOverrideCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Calendar Override Api V1 Academic Semesters  Semester Id  Calendar Overrides Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "academic"
        ],
        "summary": "List Calendar Overrides",
        "operationId": "list_calendar_overrides_api_v1_academic_semesters__semester_id__calendar_overrides_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "semester_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Semester Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Calendar Overrides Api V1 Academic Semesters  Semester Id  Calendar Overrides Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/academic/semesters/{semester_id}/calendar-overrides/{override_id}": {
      "delete": {
        "tags": [
          "academic"
        ],
        "summary": "Delete Calendar Override",
        "operationId": "delete_calendar_override_api_v1_academic_semesters__semester_id__calendar_overrides__override_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "semester_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Semester Id"
            }
          },
          {
            "name": "override_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Override Id"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SemesterCreateRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Code"
          },
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "first_monday": {
            "type": "string",
            "format": "date",
            "title": "First Monday"
          },
          "total_weeks": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 1.0,
            "title": "Total Weeks",
            "default": 20
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "start_date",
          "end_date",
          "first_monday"
        ],
        "title": "SemesterCreateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}